Make overscan crop a setting instead of using a frontend callback
In the libretro API, the client can use an environment callback to ask the
frontend if overscap should be cropped or shown. This has been moved into
a setting.

However, game.libretro will only receive settings from the
resources/settings.xml file for the particular libretro add-on being
loaded.

Therefore, if the libretro core makes use of this environment callback,
the Kodi add-on settings.xml should contain the line from
settings.xml in this commit:

<settings>
  <category label="5">
    <setting label="30000" type="bool" id="cropoverscan" default="false"/>
  </category>
</settings>